fix: acks, login, filters, and APRS_BOT_CALLSIGN
This fixes several bugs:
 * message acks were not properly removing messages from the
   _waiting_ack expiring dict, so messages were being retried that had
   been ackd. This was due to an incorrect hashing algorithim that did
   not correctly identify packets.
 * login method now checks for proper login and sends app versions and
   filters properly.

This also removes the APRS_LISTENING_CALLSIGNS that just did not work at
all. There is no mechanism in Errbot of forwarding on a "reply-as" user
for a message, so when a message for an additional callsign came in, all
replies went out as just signed in callsigns.

This is replaced by APRS_BOT_CALLSIGN, which allows setting the bot and
reply-from to a callsign separate from the one you're using to sign into
APRSIS. This is useful for running a service bot, like ANSRVR or WXBOT.
The bot will still listen on the signed in callsign as well, but if
messaged there, will repy with the APRS_BOT_CALLSIGN callsign