feat: new core plugins
This adds new core plugins for Help, Webserver, and Health and
supporting code to make it possible to load them.

I had to overload some of the Err backend methods and monkey patch some
methods in the PluginManager. Eventually, I'd like to upstream these
changes and this will no longer be needed