feat: new core plugins (#79)
* feat: new core plugins

This adds new core plugins for Help, Webserver, and Health and
supporting code to make it possible to load them.

I had to overload some of the Err backend methods and monkey patch some
methods in the PluginManager. Eventually, I'd like to upstream these
changes and this will no longer be needed

* fix: skip loading plugs if backend = True

This allows setting a new flag in the Core section of the .plug file to
indicate a plugin is a backend. These backend plugins are then skipped
when loaded by _load_plugins_generic

This code could eventually be upstreamed into Errbot to allow it to
handle multiple plugins per module easier